+++ src/components/TutorPageHead.js
@@ -23,7 +23,7 @@ function TutorPageHead({ tutor, tutorId }) {
 
   return (
     <main className="tutor-profile__profile-container">
-      <img src={tutor.profile_photo} alt="tutor profile" className="tutor-profile__photo" />
+      <img src={tutor.profilephoto} alt="tutor profile" className="tutor-profile__photo" />
       <div className="tutor-profile__profile-container-text">
         <h1 className="tutor-profile__title">{tutor.name}</h1>
         <h2 className="tutor-profile__subtitle">{tutor.profession}</h2>
