+++ src/reducers/index.js
@@ -18,7 +18,7 @@ const initialState = {
     _id: null,
     name: null,
     type: null,
-    profile_photo: null,
+    profilephoto: null,
     email: null,
     focus: null,
     description: null,
@@ -58,7 +58,7 @@ const reducer = function (state = initialState, action) {
         _id: action.payload._id,
         name: action.payload.name,
         type: action.payload.type,
-        profile_photo: action.payload.profile_photo,
+        profilephoto: action.payload.profilephoto,
         email: action.payload.email,
         focus: action.payload.focus || null,
         description: action.payload.description || null,
